src/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
gunicorn==19.9.0
celery==4.3.0
redis==3.2.1
django-cors-headers
drf-yasg
boto3
django-storages
django-health-check
django-jsonfield
psycopg2-binary
django-rest-auth
django-allauth
djangorestframework-jwt
django-rest-framework-social-oauth2